ovl/load-balancer: Use resilient nexthop for ecmp
When multipath routes are used, changes in next-hop can cause
reassignment of hash to path calculation as described in [0] and [1].

TLDR:
If a multipath group is used for load-balancing between multiple
servers, this hash space reassignment causes an issue that packets from
a single flow suddenly end up arriving at a server that does not expect
them. This can result in TCP connections being reset.

If a multipath group is used for load-balancing among available paths to
the same server, the issue is that different latencies and reordering
along the way causes the packets to arrive in the wrong order, resulting
in degraded application performance.

[0]: https://docs.kernel.org/networking/nexthop-group-resilient.html
[1]: https://netdevconf.info/0x15/papers/20/isea%20(4).pdf