ovl/test: Introduce MKROOTFS env variable
If MKROOTFS is set, xcluster_start will extend $__image with the
requested OVLs and store it in $__disk (defaults to $XCLUSTER_TMP/disk).

This will avoid having to overwrite rootfs with cdrom contents at boot
time. The default is to mkcdrom and unpack the contents into rootfs at
boot time.